fix: 탭 변경 시 Greeting 컴포넌트, '모임 만들기' 버튼 애니메이션 제외. 화면에 고정
