+++ src/components/plans/RenderCommonContent.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import { useSelector } from 'react-redux';
-import { RootState } from '@/redux/store';
+//import { useSelector } from 'react-redux';
+//import { RootState } from '@/redux/store';
 import PlanFilter from './PlanFilter';
-import CreateMeetingButton from './editMeeting/CreateMeetingButton';
+//import CreateMeetingButton from './editMeeting/CreateMeetingButton';
 import PlanList from './PlanList';
 import SortDropdown from '../shared/dropdown/SortDropdown';
 import { SortOption } from '@/types/reviewType';
@@ -41,8 +41,6 @@ const RenderCommonContent: React.FC<RenderCommonContentProps> = ({
   selectedSort,
   setSelectedSort,
 }) => {
-  const isLoggedIn = useSelector((state: RootState) => state.auth.isLoggedIn);
-
   const sortOptions: SortOption[] = [
     { id: 1, name: '최신순', value: 'default' },
     { id: 2, name: '마감임박순', value: 'closeDate' },
@@ -79,12 +77,6 @@ const RenderCommonContent: React.FC<RenderCommonContentProps> = ({
           </div>
         </div>
       </div>
-      {/* 로그인 상태일 때만 모임 만들기 버튼 표시 */}
-      {isLoggedIn && (
-        <div className="fixed bottom-20 right-5 z-10">
-          <CreateMeetingButton />
-        </div>
-      )}
       {/* 일정 카드 목록 */}
       <PlanList
         plans={plans || []}
